Add simple route
- Refactor views folder to screens
