+++ lib/features/authentication/presentation/screens/accept_invite_screen.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:pdpa/core/widgets/custom_button.dart';
+import 'package:pdpa/features/authentication/presentation/bloc/authentication_bloc.dart';
+
+class AcceptInviteScreen extends StatelessWidget {
+  const AcceptInviteScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Theme.of(context).colorScheme.onPrimary,
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: <Widget>[
+          CustomButton(
+            height: 40.0,
+            onPressed: () {
+              context.read<AuthenticationBloc>().add(const SignOutEvent());
+            },
+            buttonType: CustomButtonType.outlined,
+            child: Text(
+              'Sign out',
+              style: Theme.of(context)
+                  .textTheme
+                  .titleMedium
+                  ?.copyWith(color: Theme.of(context).colorScheme.primary),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
